:pencil2: fixed typo in script name :muscle: finished all count functions :contruction: rtemp output a table, will work on outputting vcf
